pyproject.toml
[tool.poetry]
name = "nonebot-plugin-gspanel"
version = "0.2.13"
description = "Genshin player cards plugin for NoneBot2"
authors = ["monsterxcn <monsterxcn@gmail.com>"]
documentation = "https://github.com/monsterxcn/nonebot-plugin-gspanel#readme"
license = "MIT"
homepage = "https://github.com/monsterxcn/nonebot-plugin-gspanel"
readme = "README.md"
keywords = ["nonebot", "nonebot2", "genshin", "panel", "card"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
nonebot2 = ">=2.0.0b3"
nonebot-adapter-onebot = ">=2.0.0b1"
httpx = ">=0.20.0, <1.0.0"
nonebot-plugin-htmlrender = ">=0.2.0.1"
playwright = ">=1.25.0"

[tool.poetry.group.dev.dependencies]
pycln = "^2.1.2"
pyupgrade = "^3.3.0"
black = "^22.10.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
pre-commit = "^2.19.0"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.pycln]
path = "."
all = false

[tool.pyright]
reportShadowedImports = false
pythonVersion = "3.8"
pythonPlatform = "All"
executionEnvironments = [
  { root = "./tests", extraPaths = ["./"] },
  { root = "./" },
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
